structural changes to locationdrawings
